+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ next version
 * Add modeling.get_unit_of
 * Add modeling.get_concentration_parameters_from_data
 * Add modeling.write_csv
+* Add modeling.resample_data
+* Add modeling.omit_data
 * Support for Python 3.7 dropped
 
 0.41.0 (2021-12-21)